+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- species concentration at the mouseover location in the species tab [#122](https://github.com/spatial-model-editor/spatial-model-editor/issues/122)
+
 ### Fixed
 - ImageSlice dialog now uses the currently selected z-slice, mouseover text reports physical `x/y/z/t` values, geometry image has grid and scale overlays [#577](https://github.com/spatial-model-editor/spatial-model-editor/issues/577)
+- exported TIFF files now include geometry origin [#412](https://github.com/spatial-model-editor/spatial-model-editor/issues/412)
 
 ## [1.12.0] - 2026-04-22
 ### Added